examples/patterns/basic-chat/python/requirements.txt
# nxuskit - install from local package or PyPI when available
# pip install -e ../../packages/nxuskit-py
nxuskit-py
